8.sqlalchemy/3.sql_alchemy-basics/requirements.txt
alembic==1.16.5
Faker==37.8.0
greenlet==3.2.4
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.2
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzdata==2025.2
